New name of main.cpp and added documentation files
