+++ docs/README.md
@@ -1,0 +1,8 @@
+# Docs
+
+## Navigation
+
+Our GitHub documentation consist of:
+
+- [Metric Fetching Functions](./metric-fetching-functions.md)
+- [Table of Metrics](./metrics-description.md)
